completed todolist

+++ app/Http/Controllers/NoteController.php
@@ -12,8 +12,26 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderby('created_at', 'desc')->paginate();
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderby('created_at', 'desc')
+            ->paginate();
+
         return view('note.index', ['notes' => $notes]);
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     */
+    public function store(Request $request)
+    {
+        $data = $request->validate([
+            'note' => ['required', 'string'],
+        ]);
+        $data['user_id'] = $request->user()->id;
+        $note = Note::create($data);
+
+        return to_route('note.show', $note)->with('message', 'Note created successfully');
     }
 
     /**
@@ -25,23 +43,13 @@ class NoteController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        $data = $request->validate([
-            'note' => ['required', 'string'],
-        ]);
-        $data['user_id'] = 1;
-        $note = Note::create($data);
-        return to_route('note.show', $note)->with('message', 'Note created successfully');
-    }
-
-    /**
      * Display the specified resource.
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -50,6 +58,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -58,10 +69,15 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $data = $request->validate([
             'note' => ['required', 'string'],
         ]);
         $note->update($data);
+
         return to_route('note.show', $note)->with('message', 'Note updated successfully');
     }
 
@@ -70,7 +86,11 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
+
         return to_route('note.index')->with('message', 'Note deleted successfully');
     }
 }
